New Backgrounds, Extra Story
There are now backgrounds for mills, snoddy center, study corral, and couch room. They are all now shown in the game at least in one place, we can implement this further in the future. I have also focused on writing some more of the "coming soon" story. We are simplifying where possible for the time being. Also, everything is now cited.